Fixed backend to enable CORS. Everything deployed!
