+++ guitar-api/src/main/java/com/kermatt/guitar_api/GuitarInfoApiApplication.java
@@ -2,6 +2,9 @@ package com.kermatt.guitar_api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @SpringBootApplication
 public class GuitarInfoApiApplication {
@@ -10,4 +13,17 @@ public class GuitarInfoApiApplication {
 		SpringApplication.run(GuitarInfoApiApplication.class, args);
 	}
 
+	@Bean
+    public WebMvcConfigurer corsConfig() {
+        System.out.println("Adding CORS configuration pt1...");
+        return new WebMvcConfigurer() {
+            @Override
+            public void addCorsMappings(CorsRegistry registry) {
+                registry.addMapping("/**")
+                    .allowedOrigins("https://d3v8d7zaba9elv.cloudfront.net", "http://localhost:4200")
+                    .allowedMethods("GET")
+                    .allowedHeaders("content-type", "authorization");
+            }
+        };
+    }
 }
