+++ guitar-api/src/main/resources/application.properties
@@ -1,14 +1,8 @@
-# spring.application.name=Guitar Info API
-# spring.datasource.url=jdbc:postgresql://localhost/postgres
-# spring.datasource.username=${PSQL_USER}
-# spring.datasource.password=${PSQL_PASSWORD}
-# spring.datasource.baseUrl=http://localhost:8080/api/v1
-# spring.jpa.hibernate.ddl-auto=update
-# spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 spring.application.name=Guitar Info API
-spring.datasource.url=jdbc:postgresql://guitar-database-instance.cnas644guogg.us-east-2.rds.amazonaws.com:5432/guitar_db
+# spring.datasource.url=jdbc:postgresql://localhost/postgres
 spring.datasource.username=${PSQL_USER}
 spring.datasource.password=${PSQL_PASSWORD}
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
-logging.level.org.springframework.web=DEBUG
+spring.datasource.url=jdbc:postgresql://guitar-database-instance.cnas644guogg.us-east-2.rds.amazonaws.com:5432/guitar_db
+# spring.datasource.baseUrl=http://localhost:8080/api/v1
